requirements.txt
alembic==1.0.11
certifi==2019.6.16
chardet==3.0.4
distro==1.4.0
idna==2.8
lxml==4.3.4
Mako==1.0.13
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
PyPDF2==1.26.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
scraperwiki==0.5.1
six==1.12.0
SQLAlchemy==1.3.5
tabula-py==1.3.1
urllib3==1.25.3
